+++ tools/azsdk-cli/Azure.Sdk.Tools.Cli/Commands/SharedOptions.cs
@@ -26,7 +26,7 @@ namespace Azure.Sdk.Tools.Cli.Commands
             typeof(TelemetryIngestionTool),
             typeof(PackageInfoTool),
             typeof(PackageCheckTool),
-            typeof(PackageWorkItemLookupTool),
+            typeof(PackageWorkItemTool),
             typeof(PipelineTestsTool),
             typeof(QuokkaTool),
             typeof(ReadMeGeneratorTool),